fix: resolve 34 failing tests across Excel, DOCX, PPTX, and Chart handlers
Excel: add rich text runs, conditional formatting types (topN, aboveAverage,
uniqueValues, duplicateValues, containsText, dateOccurring), cell links,
array formulas, page break removal, and formula-after-shared-string readback.

DOCX: fix inline SDT query paths (/body/p[N]/sdt[M]), field type dispatch
via "type" property fallback, and oMathPara ElementToNode mapping.

PPTX: copy hyperlink/external relationships on cross-slide move, skip
default P14 transition direction suffix, normalize alignment values.

Chart: separate gridline presence flag from detail keys (gridlineColor,
gridlineWidth, gridlineDash).

+++ src/officecli/Handlers/Excel/ExcelHandler.Query.cs
@@ -227,7 +227,7 @@ public partial class ExcelHandler
             return new DocumentNode
             {
                 Path = path, Type = "rowbreak",
-                Format = { ["row"] = (int)(brk.Id?.Value ?? 0u), ["manual"] = brk.ManualPageBreak?.Value ?? false }
+                Format = { ["row"] = brk.Id?.Value ?? 0u, ["manual"] = brk.ManualPageBreak?.Value ?? false }
             };
         }
         var cbMatch = Regex.Match(cellRef, @"^colbreak\[(\d+)\]$", RegexOptions.IgnoreCase);
